requirements.txt
selfbotUtils==0.0.2
discord.py-self